+++ erui-parent/erui-order/src/main/java/com/erui/order/service/impl/ProjectServiceImpl.java
@@ -59,8 +59,8 @@ public class ProjectServiceImpl implements ProjectService {
     @Override
     public boolean updateProject(Project project) throws Exception {
         Project projectUpdate = projectDao.findOne(project.getId());
-        Long proNameId = projectDao.findIdByProjectName(project.getProjectName());
-        if (proNameId != null && proNameId.intValue() != project.getId()) {
+        Integer proNameId = projectDao.findIdByProjectName(project.getProjectName());
+        if (proNameId!= null && proNameId.intValue() != project.getId()) {
             throw new Exception("项目名称已存在");
         }
         if (project.getProjectStatus().equals("SUBMIT") || project.getProjectStatus().equals("HASMANAGER") || project.getProjectStatus().equals("EXECUTING")) {
